updated schema validation
